# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "YSpotify"
    },
    "paths": {
        "/signup": {
            "get": {
                "summary": "FT-1 Inscription",
                "description": "Créer un compte",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie"
                    }
                }
            }
        },
        "/signin": {
            "get": {
                "summary": "FT-2 Connexion",
                "description": "Pour s'authentifier",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie"
                    }
                }
            }
        },
        "/auth-url": {
            "get": {
                "summary": "FT-3 Liaison du compte Spotify",
                "description": "Pour s'authentifier avec son compte Spotify",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token JWT",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie"
                    }
                }
            }
        },
        "/group": {
            "get": {
                "summary": "FT-4 Rejoindre un Groupe",
                "description": "Changer de groupe",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token JWT",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nom du groupe",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie"
                    }
                }
            }
        },
        "/grouplist": {
            "get": {
                "summary": "FT-5 Consultation des Groupes et Utilisateurs - Liste des Groupes",
                "description": "Lister l'ensemble des groupes",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token JWT",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie"
                    }
                }
            }
        },
        "/mygroup": {
            "get": {
                "summary": "FT-5 Consultation des Groupes et Utilisateurs - Liste des membres du Groupe",
                "description": "Afficher les utilisateurs de mon groupe",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token JWT",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie"
                    }
                }
            }
        },
        "/userpersonality": {
            "get": {
                "summary": "FT-6 Personnalité de l’Utilisateur",
                "description": "Afficher la personalité de l'utilisateur",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token JWT",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie"
                    }
                }
            }
        },
        "/sync": {
            "get": {
                "summary": "FT-7 Synchronisation",
                "description": "Si l'utilisateur est le chef du groupe, il peut synchroniser la musique qu’il est en train d’écouter sur tous les appareils actifs des autres Utilisateurs ",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token JWT",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie"
                    }
                }
            }
        },
        "/addplaylist": {
            "post": {
                "summary": "FT-8 Playlist",
                "description": "L’Utilisateur peut demander la création d’une playlist sur son compte Spotify contenant les 10 musiques préférées d’un autre Utilisateur (qui peut être lui-même) passé en paramètre",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token JWT",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Nom de l'utilisateur",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "nom",
                        "in": "query",
                        "description": "Nom de la playlist",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "des",
                        "in": "query",
                        "description": "Description de la playlist",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie"
                    }
                }
            }
        }
    }
}